+++ Water_Features/Properties/Stable/ChangeLog.md
@@ -2,3 +2,4 @@
 * While in-game and using the modern water simulation, adds a field for seeing Sea Level that can be modified with a configurable slider and text field input.
 * Compatibility fixes with game version 1.5.2f1.
 * Update localization.
+* When clicking Upgrade to New Water Simulation in DevUI Menu, will update water source prefabs for Yenyang's Custom Water Tool.